requirements.txt
google-genai>=1.0.0
click>=8.0.0
rich>=13.0.0
